240723 change the default app on the flutter
